+++ src/extension/boostConfiguration.ts
@@ -104,6 +104,15 @@ export class BoostConfiguration {
             .update(Defaults.addBoostToAnalyzedProjectsName, value, ConfigurationTarget.Global);
     }
 
+    public static get cleanupUnusedBoostFilesAutomatically(): boolean {
+        return (workspace.getConfiguration(NOTEBOOK_TYPE, null).get(Defaults.cleanupUnusedBoostFilesAutomaticallyName) as boolean)??
+            Defaults.cleanupUnusedBoostFilesAutomaticallyValue;
+    }
+    public static set cleanupUnusedBoostFilesAutomatically(value: string) {
+        workspace.getConfiguration(NOTEBOOK_TYPE, null)
+            .update(Defaults.cleanupUnusedBoostFilesAutomaticallyName, value, ConfigurationTarget.Global);
+    }
+
     public static get logLevel(): string {
         return (workspace.getConfiguration(NOTEBOOK_TYPE, null).get(Defaults.logLevelName) as string)??
             Defaults.logLevelValue;
@@ -266,6 +275,9 @@ class Defaults {
 
     public static readonly addBoostToAnalyzedProjectsName : string = "advanced.addBoostToAnalyzedProjects";
     public static readonly addBoostToAnalyzedProjectsValue : boolean = true;
+
+    public static readonly cleanupUnusedBoostFilesAutomaticallyName : string = "advanced.cleanupUnusedBoostFilesAutomatically";
+    public static readonly cleanupUnusedBoostFilesAutomaticallyValue : boolean = true;
 
     // enables targeting a specific kernel for analysis when using run all in UX
     // for example, if you want to run only security analysis with source id and no summary, use: